+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cat-factory/eks
 
+## 0.1.267
+
+### Patch Changes
+
+- Updated dependencies [55310f6]
+- Updated dependencies [55310f6]
+  - @cat-factory/contracts@0.271.0
+  - @cat-factory/kernel@0.269.0
+  - @cat-factory/integrations@0.147.0
+
 ## 0.1.266
 
 ### Patch Changes