+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/provider-s3
 
+## 0.2.336
+
+### Patch Changes
+
+- Updated dependencies [55310f6]
+- Updated dependencies [55310f6]
+  - @cat-factory/kernel@0.269.0
+
 ## 0.2.335
 
 ### Patch Changes